+++ stories/with-configuration.stories.js
