+++ stories/with-no-timer.stories.js
@@ -11,30 +11,9 @@ export default {
   },
 };
 
-export function Default({ bodyText }) {
-  return (
-    <ToastProvider clientKey="client-key">
-      <DefaultComponent body={bodyText} />
-    </ToastProvider>
-  );
-}
-
 function DefaultComponent({ body }) {
   const [toast] = useToast();
   return <Button onClick={() => toast(body)}>Show Toast</Button>;
-}
-
-export function NoIcon({ bodyText }) {
-  return (
-    <ToastProvider
-      clientKey="client-key"
-      config={{
-        icon: false,
-      }}
-    >
-      <DefaultComponent body={bodyText} />
-    </ToastProvider>
-  );
 }
 
 export function NoTimer({ bodyText }) {
